refactor: renamed methods
renamed methods fetchMarketApps, findApp, installApp, fetchStoreApplications, findStoreApplication,
editApplication and removeApplication

BREAKING CHANGE: renamed fetchMarketApps => listFromMarket
 findApp => findOnMarket
 installApp
=> install
 fetchStoreApplications => list
 findStoreApplication => find
 editApplication =>
edit
 removeApplication => remove

"re #58", "fix #57".
