+++ src/methods/remove.js
